Version with globals

+++ client.c
@@ -6,7 +6,7 @@
 /*   By: amohiam <amohiam@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/02 15:03:49 by amohiam           #+#    #+#             */
-/*   Updated: 2021/09/05 15:24:21 by amohiam          ###   ########.fr       */
+/*   Updated: 2021/09/21 15:15:54 by amohiam          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,12 +18,12 @@ void	send_bits(int pid, char *message)
 	int	i;
 
 	i = 0;
-	while (i < ft_strlen(message))
+	while (i < (int)ft_strlen(message))
 	{
 		bit_num = 0;
 		while (bit_num < 7)
 		{
-			if((message[i] >> bit_num) & 1)
+			if ((message[i] >> bit_num) & 1)
 				kill(pid, SIGUSR2);
 			else
 				kill(pid, SIGUSR1);
@@ -34,9 +34,10 @@ void	send_bits(int pid, char *message)
 	}
 }
 
-int		main(int argc, char **argv)
+int	main(int argc, char **argv)
 {
 	int	pid;
+
 	if (argc == 3)
 	{
 		pid = ft_atoi(argv[1]);
